backend/requirements.txt
fastapi
alembic
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
httpx
firebase-admin
python-jose
passlib[bcrypt]
pydantic
pydantic-settings
email-validator
python-multipart
numpy
spacy
sentence-transformers
langchain
huggingface_hub
mistralai
langchain-mistralai
langchain-community
langchain-chroma
chromadb
pandas
python-slugify
Pillow
beautifulsoup4

